add condition on delete todo if id undefinied

+++ src/Handlers/Todos/DeleteTodo.js
@@ -2,15 +2,17 @@ import React from 'react'
 import axios from 'axios'
 import { DELETE_TODOS_URL } from '../../API/constants'
 const DeleteTodo = async (id, setLoading, setError)=> {
+  if(id){
     setLoading(true)
     try {
-      await axios.delete(`${DELETE_TODOS_URL}/${id}`)
-      setLoading(false)
-      setError(false)
-    } catch (error) {
-      setLoading(false)
-      setError(true)
+        await axios.delete(`${DELETE_TODOS_URL}/${id}`)
+        setLoading(false)
+        setError(false)
+      } catch (error) {
+        setLoading(false)
+        setError(true)
     }
+  }
 }
 
 export default DeleteTodo
